Split progress merge tests
